Move search logic out of Storage class
- new search-index-new/search feature module to work on search
